+++ fractals/src/main/java/me/catmousedog/fractals/fractals/concrete_fractals/Mandelbrot.java
@@ -41,7 +41,7 @@ public class Mandelbrot extends Fractal {
 			y = 2 * tx * y + cy;
 
 		}
-		return new FractalValue();
+		return new FractalValue(0, 0, iterations, iterations);
 	}
 
 	public @NotNull String informalName() {